Refactoring into different components
