+++ addon/components/universal-export/template.hbs
@@ -26,32 +26,4 @@
   {{/if}}
 {{/modal-view}}
 
-
-{{#modal-view hidden=hideInfluencerNetwork closeAction=closeInfluencerNetworkModal centered=true
-  borderlessHeader=true title="Influencer Network" container="body"}}
-  <div class="modal-body text-center">
-    <p class="text-size-7">
-      <b>Influencer has been added to your network</b>
-    </p>
-    <p class="text-size-5 text-color-default-lighter padding-left-md padding-right-md">
-      By adding this influencer to your list, it was automatically added to your Influencers Network. You can access and edit your network in your IRM.
-    </p>
-
-    <div>
-      <a {{action "closeInfluencerNetworkModal"}} target="_blank" class="upf-btn upf-btn--default margin-top-x-sm margin-bottom-x-sm margin-right-xx-sm">
-        Skip
-      </a>
-      <a href="http://help.upfluence.co" target="_blank" class="upf-btn upf-btn--primary margin-top-x-sm margin-bottom-x-sm">
-        Learn More
-      </a>
-    </div>
-
-    <div>
-    {{#input-wrapper class="influencer-network-checkbox"}}
-      {{#upf-checkbox value=hasDisabledInfluencerNetworkModal hasLabel=true onValueChange=(action "dontShowInfluencerNetworkModal")}}
-        <span>Don't show me this again</span>
-      {{/upf-checkbox}}
-    {{/input-wrapper}}
-    </div>
-  </div>
-{{/modal-view}}
+{{influencer-network-modal hideInfluencerNetwork=hideInfluencerNetwork}}
