+++ app/components/influencer-network-modal/component.js
@@ -1,0 +1,1 @@
+export { default } from '@upfluence/ember-upf-utils/components/influencer-network-modal/component';
